Push materialisation mutex deep into materialiser; Fix IAM benchmark build (#3)
## What is the goal of this PR?
The move the `synchronized` construct from the materialise method of ConclusionNode deep into the materialiser. It is now present only on the relation materialisation branch and is synchronised on the type of the relation.

We also introduce a flag enabling "Relaxed relation inference semantics" - An relation will be inferred even if a persisted relation that subsumes it exists.  This is not enabled by default.